Added the game difficulty configuration class and added the ability to generate scenes after projectiles kill enemies. 加入游戏难度的配置类，加入抛射体杀死敌人后生成场景的功能。

+++ scripts/loot/LootEntry.cs
@@ -47,7 +47,7 @@ public readonly struct LootEntry(string itemId, int minQuantity = 1, int maxQuan
 ///<para>Entries</para>
 ///<para>条目列表</para>
 /// </param>
-public readonly record struct LootGroup(double Chance, IEnumerable<LootEntry> Entries)
+public readonly record struct LootGroup(float Chance, IEnumerable<LootEntry> Entries)
 {
     private int WeightSum { get; } = Entries.Sum(entry => entry.Weight);
